list filter

+++ src/components/MainPage.vue
@@ -1,11 +1,11 @@
 <template>
 
-  <div>
+
 
     <div class="row">
       <div class="col-md-8">
         <div class="detail" v-if="selectedEntry">
-          <EntryDetail :entry="selectedEntry"></EntryDetail>
+          <EntryDetail :entry="selectedEntry" ></EntryDetail>
           <div  class="ui buttons container">
             <button class="teal ui button" @click="toggleUpdate">Update</button>
             <button class="orange ui button" @click="deleteEntry(selectedEntry.id)">Delete</button>
@@ -15,32 +15,26 @@
 
         <div v-if="update" class="container update">
           <div class="mb-3">
+            <h3>Update</h3>
             <label class="form-label">Headline</label>
             <input type="text" class="form-control" :placeholder="selectedEntry.headline" id="updateHeadline">
           </div>
           <div class="mb-3">
             <label  class="form-label">Entry</label>
-            <textarea class="form-control" rows="4" :placeholder="selectedEntry.entry" id="updateEntry"></textarea>
+            <textarea class="form-control" rows="4" id="updateEntry" :placeholder="selectedEntry.entry" ></textarea>
           </div>
           <button class="ui olive button" @click="updateEntry(selectedEntry.id)">Confirm</button>
         </div>
 
       </div>
-
-
       <div class="col-md-4">
 
         <EntryList :entries="entries"
                    @entrySelect="onEntrySelect"></EntryList>
       </div>
-
-
-
-
-
     </div>
 
-  </div>
+
 </template>
 <script>
 
@@ -60,9 +54,12 @@ export default {
     return {
       entries: [],
       selectedEntry: null,
-      update: false
+      selectedId: null,
+      update: false,
+
     }
   },
+
   methods: {
     fetchEntries() {
       axios.get(`${ROOT_URL}/entries`)
@@ -72,61 +69,71 @@ export default {
           }
       );
     },
+
     deleteEntry(id) {
-      return axios.delete(`${ROOT_URL}/entries/` + id)
-          .then(
-              this.fetchEntries,
-              this.selectedEntry = null
-          )
-
+      if(window.confirm("are you sure about deleting?")) {
+        this.removeUpdate()
+        this.removeSelected()
+        return axios.delete(`${ROOT_URL}/entries/` + id)
+            .then(
+                this.fetchEntries,
+            )
+      }
     },
+
     updateEntry(id){
-      const headline = document.getElementById("updateHeadline").value;
-      const entry = document.getElementById("updateEntry").value;
+      if (window.confirm("are you sure about updating?")) {
 
-      axios.put(`${ROOT_URL}/entries/` + id, {
-        headline, entry
-      }).then(
-          this.fetchEntries,
-      )
+        const headline = document.getElementById("updateHeadline").value;
+        const entry = document.getElementById("updateEntry").value;
 
+        this.removeUpdate()
+        this.removeSelected()
+        axios.put(`${ROOT_URL}/entries/` + id, {
+          headline, entry
+        }).then(
+            this.fetchEntries
+        )
+      }
     },
     onEntrySelect(entry) {
       this.selectedEntry = entry;
     },
     toggleUpdate() {
       this.update = !this.update
+    },
+    removeSelected() {
+      this.selectedEntry = null;
+    },
+    removeUpdate() {
+      if(this.update) this.update = false;
     }
   },
+
   created() {
     this.fetchEntries()
   },
-
 }
 </script>
 
+
 <style scoped>
-
 .update{
-
   padding: 10px;
   border: 1px solid dimgray;
-  border-radius: 4px;
+  border-radius: 15px;
+  background-color: aliceblue;
 
 }
 
 .buttons{
-
   padding: 10px;
-
 }
 
 .detail{
-
   padding: 10px;
   border: 1px solid dimgray;
-  border-radius: 4px;
-
+  border-radius: 15px;
+  background-color: aliceblue;
 }
-
 </style>
